# openapi.json
{"openapi":"3.0.0","paths":{"/sessions":{"post":{"operationId":"createSession","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateQuizSession"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizSessionSecure"}}}}},"tags":["Session"]}},"/sessions/resolve":{"post":{"operationId":"resolveSession","parameters":[{"name":"x-session-key","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizSessionPublic"}}}}},"tags":["Session"]}},"/sessions/answers/{index}":{"put":{"operationId":"setSessionAnswer","parameters":[{"name":"x-session-key","required":true,"in":"header","schema":{"type":"string"}},{"name":"index","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateQuestionAnswer"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizSessionPublic"}}}}},"tags":["Session"]}},"/sessions/finish":{"post":{"operationId":"finishSession","parameters":[{"name":"x-session-key","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizSessionPublic"}}}}},"tags":["Session"]}},"/quiz":{"get":{"operationId":"listQuizes","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuizPreview"}}}}}},"tags":["Quiz"]}},"/quiz/{index}":{"get":{"operationId":"getQuiz","parameters":[{"name":"index","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizPublic"}}}},"404":{"description":"Quiz not found"}},"tags":["Quiz"]}}},"info":{"title":"Cats example","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateQuizSession":{"type":"object","properties":{"quizId":{"type":"number"}},"required":["quizId"]},"QuizSessionSecure":{"type":"object","properties":{"quizId":{"type":"number"},"status":{"type":"string","enum":["in-progress","finished"]},"answers":{"type":"array","items":{"type":"array","items":{"type":"number"}}},"feedback":{"type":"object","properties":{"questions":{"type":"array","items":{"type":"object","properties":{"currentAnswer":{"type":"array","items":{"type":"number"}},"rightAnswer":{"type":"array","items":{"type":"number"}},"comment":{"type":"string"},"score":{"type":"number"},"scoreMax":{"type":"number"}},"required":["currentAnswer","rightAnswer","comment","score","scoreMax"],"additionalProperties":false}},"score":{"type":"number"},"scoreMax":{"type":"number"}},"required":["questions","score","scoreMax"],"additionalProperties":false},"key":{"type":"string"}},"required":["quizId","status","answers","key"]},"QuizSessionPublic":{"type":"object","properties":{"quizId":{"type":"number"},"status":{"type":"string","enum":["in-progress","finished"]},"answers":{"type":"array","items":{"type":"array","items":{"type":"number"}}},"feedback":{"type":"object","properties":{"questions":{"type":"array","items":{"type":"object","properties":{"currentAnswer":{"type":"array","items":{"type":"number"}},"rightAnswer":{"type":"array","items":{"type":"number"}},"comment":{"type":"string"},"score":{"type":"number"},"scoreMax":{"type":"number"}},"required":["currentAnswer","rightAnswer","comment","score","scoreMax"],"additionalProperties":false}},"score":{"type":"number"},"scoreMax":{"type":"number"}},"required":["questions","score","scoreMax"],"additionalProperties":false}},"required":["quizId","status","answers"]},"UpdateQuestionAnswer":{"$schema":"https://json-schema.org/draft/2020-12/schema","type":"array","items":{"type":"number"}},"QuizPreview":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}},"required":["title","description"]},"QuizPublic":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"questions":{"type":"array","items":{"type":"object","properties":{"description":{"type":"string"},"options":{"type":"array","items":{"type":"object","properties":{"description":{"type":"string"}},"required":["description"],"additionalProperties":false}},"mode":{"type":"string","enum":["single","multiple"]},"score":{"type":"number"},"answer":{"type":"array","items":{"type":"number"}}},"required":["description","options","mode","score","answer"],"additionalProperties":false}}},"required":["title","description","questions"]}}}}